Refactored authentication service
