+++ core/api/auth/package.json
@@ -16,6 +16,7 @@
         "body-parser": "^1.19.0",
         "cors": "^2.8.5",
         "express": "^4.17.1",
+        "helmet": "^4.4.0",
         "jsonwebtoken": "^8.5.1",
         "pg": "^8.5.1"
     },
@@ -25,7 +26,6 @@
         "@types/body-parser": "^1.19.0",
         "@types/cors": "^2.8.9",
         "@types/express": "^4.17.9",
-        "@types/pg": "^7.14.7",
         "nodemon": "^2.0.7",
         "ts-node": "^9.1.1",
         "typescript": "^4.1.3"
