+++ core/api/auth/src/db.ts
@@ -1,9 +1,10 @@
-import { PoolConfig } from 'pg';
+import { Pool, PoolConfig } from 'pg';
 
-export const dbConfig: PoolConfig = {
+
+export const pool: Pool = new Pool({
     host: 'userdb',
     port: 5432,
     user: process.env.POSTGRES_USER,
     password: process.env.POSTGRES_PASSWORD,
     database: process.env.POSTGRES_DB 
-};
+});
