+++ core/api/auth/src/types/User.ts
@@ -1,0 +1,10 @@
+type User = {
+    userid: number,
+    handle: string,
+    email: string,
+    password: string,
+    firstname: string,
+    lastname: string,
+};
+
+export default User
