+++ core/api/auth/src/utils/jwtAuth.ts
@@ -1,12 +1,5 @@
 import jwt from "jsonwebtoken";
-
-type User = {
-    handle: string,
-    email: string,
-    firstname: string,
-    lastname: string,
-    password: string
-}
+import User from "../types/User";
 
 export const generateToken = (user: User): string => {
     return jwt.sign(user, process.env.JWT_SECRET, {
